requirements.txt
google-adk
python-dotenv
google-generativeai
uvicorn
requests
fastapi
pydantic
googlesearch-python

# Testing dependencies
pytest
pytest-asyncio
pytest-mock
httpx
